+++ packages/amplify-graphql-index-transformer/src/resolvers.ts
@@ -60,7 +60,7 @@ const API_KEY = 'API Key Authorization';
  * replaceDdbPrimaryKey
  */
 export function replaceDdbPrimaryKey(config: PrimaryKeyDirectiveConfiguration, ctx: TransformerContextProvider): void {
-  // Replace the table's primary key with the value from @primaryKey.
+  // Replace the table's primary key with the value from @primaryKey
   const { field, object } = config;
   const table = getTable(ctx, object) as any;
   const cfnTable = table.table;
@@ -507,6 +507,10 @@ function makeQueryResolver(config: IndexDirectiveConfiguration, ctx: Transformer
               set(
                 ref('filterExpression'),
                 methodCall(ref('util.parseJson'), methodCall(ref('util.transform.toDynamoDBFilterExpression'), ref('filter'))),
+              ),
+              iff(
+                isNullOrEmpty(ref('filterExpression')),
+                methodCall(ref('util.error'), str('Unable to process the filter expression'), str('Unrecognized Filter')),
               ),
               iff(
                 not(methodCall(ref('util.isNullOrBlank'), ref('filterExpression.expression'))),