+++ packages/amplify-util-mock/src/__e2e_v2__/model-relational-with-key-transformer.e2e.test.ts
@@ -1,11 +1,11 @@
-import { ModelTransformer } from '@aws-amplify/graphql-model-transformer';
-import { HasManyTransformer, BelongsToTransformer, HasOneTransformer } from '@aws-amplify/graphql-relational-transformer';
-import { PrimaryKeyTransformer, IndexTransformer } from '@aws-amplify/graphql-index-transformer';
-import { GraphQLTransform } from '@aws-amplify/graphql-transformer-core';
-import { FeatureFlagProvider } from '@aws-amplify/graphql-transformer-interfaces';
-import { deploy, launchDDBLocal, terminateDDB, logDebug, GraphQLClient } from '../__e2e__/utils';
-import { AuthTransformer } from '@aws-amplify/graphql-auth-transformer';
-import { AmplifyAppSyncSimulator } from 'amplify-appsync-simulator';
+import { ModelTransformer } from "@aws-amplify/graphql-model-transformer";
+import { HasManyTransformer, BelongsToTransformer, HasOneTransformer } from "@aws-amplify/graphql-relational-transformer";
+import { PrimaryKeyTransformer, IndexTransformer } from "@aws-amplify/graphql-index-transformer";
+import { GraphQLTransform } from "@aws-amplify/graphql-transformer-core";
+import { FeatureFlagProvider } from "@aws-amplify/graphql-transformer-interfaces";
+import { AuthTransformer } from "@aws-amplify/graphql-auth-transformer";
+import { AmplifyAppSyncSimulator } from "@aws-amplify/amplify-appsync-simulator";
+import { deploy, launchDDBLocal, terminateDDB, logDebug, GraphQLClient } from "../__e2e__/utils";
 
 let GRAPHQL_CLIENT: GraphQLClient;
 let GRAPHQL_ENDPOINT: string;
@@ -15,7 +15,7 @@ let server: AmplifyAppSyncSimulator;
 
 jest.setTimeout(20000);
 
-describe('@model with relational transformers', () => {
+describe("@model with relational transformers", () => {
   beforeAll(async () => {
     const validSchema = `
       type AProject @model(subscriptions: null) @auth(rules: [{ allow: public }]) {
@@ -73,21 +73,21 @@ describe('@model with relational transformers', () => {
           new HasOneTransformer(),
           new HasManyTransformer(),
           new BelongsToTransformer(),
-          new AuthTransformer(),
+          new AuthTransformer()
         ],
-        featureFlags: ({
+        featureFlags: {
           getBoolean: (featureName: string, defaultValue: boolean) => {
-            if (featureName === 'improvePluralization') {
+            if (featureName === "improvePluralization") {
               return true;
             }
 
-            if (featureName === 'respectPrimaryKeyAttributesOnConnectionField') {
+            if (featureName === "respectPrimaryKeyAttributesOnConnectionField") {
               return false;
             }
 
             return defaultValue;
-          },
-        } as FeatureFlagProvider),
+          }
+        } as FeatureFlagProvider
       });
       const out = transformer.transform(validSchema);
 
@@ -96,13 +96,13 @@ describe('@model with relational transformers', () => {
       const result = await deploy(out, ddbClient);
       server = result.simulator;
 
-      GRAPHQL_ENDPOINT = server.url + '/graphql';
+      GRAPHQL_ENDPOINT = `${server.url}/graphql`;
       logDebug(`Using graphql url: ${GRAPHQL_ENDPOINT}`);
 
-      const apiKey = result.config.appSync.apiKey;
+      const { apiKey } = result.config.appSync;
       logDebug(apiKey);
       GRAPHQL_CLIENT = new GraphQLClient(GRAPHQL_ENDPOINT, {
-        'x-api-key': apiKey,
+        "x-api-key": apiKey
       });
     } catch (e) {
       console.error(e);
@@ -126,7 +126,7 @@ describe('@model with relational transformers', () => {
    * Test queries below
    */
 
-  test('Unnamed connection 1 way navigation, with @primaryKey directive 1:1', async () => {
+  test("Unnamed connection 1 way navigation, with @primaryKey directive 1:1", async () => {
     await GRAPHQL_CLIENT.query(
       `
       mutation CreateATeam {
@@ -136,7 +136,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     await GRAPHQL_CLIENT.query(
@@ -148,7 +148,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     const queryResponse = await GRAPHQL_CLIENT.query(
@@ -166,17 +166,17 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
     expect(queryResponse.data.listAProjects).toBeDefined();
-    const items = queryResponse.data.listAProjects.items;
+    const { items } = queryResponse.data.listAProjects;
     expect(items.length).toEqual(1);
-    expect(items[0].projectId).toEqual('P1');
+    expect(items[0].projectId).toEqual("P1");
     expect(items[0].team).toBeDefined();
-    expect(items[0].team.teamId).toEqual('T1');
+    expect(items[0].team.teamId).toEqual("T1");
   });
 
-  test('Unnamed connection 1 way navigation, with @primaryKey directive 1:M', async () => {
+  test("Unnamed connection 1 way navigation, with @primaryKey directive 1:M", async () => {
     await GRAPHQL_CLIENT.query(
       `
       mutation CreateBProject {
@@ -186,7 +186,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     await GRAPHQL_CLIENT.query(
@@ -198,7 +198,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     await GRAPHQL_CLIENT.query(
@@ -210,7 +210,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     const queryResponse = await GRAPHQL_CLIENT.query(
@@ -230,20 +230,20 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
     expect(queryResponse.data.listBProjects).toBeDefined();
-    const items = queryResponse.data.listBProjects.items;
+    const { items } = queryResponse.data.listBProjects;
     expect(items.length).toEqual(1);
-    expect(items[0].projectId).toEqual('P1');
+    expect(items[0].projectId).toEqual("P1");
     expect(items[0].teams).toBeDefined();
     expect(items[0].teams.items).toBeDefined();
     expect(items[0].teams.items.length).toEqual(2);
-    expect(items[0].teams.items[0].teamId).toEqual('T1');
-    expect(items[0].teams.items[1].teamId).toEqual('T2');
+    expect(items[0].teams.items[0].teamId).toEqual("T1");
+    expect(items[0].teams.items[1].teamId).toEqual("T2");
   });
 
-  test('Named connection 2 way navigation, with with custom @primaryKey fields 1:1', async () => {
+  test("Named connection 2 way navigation, with with custom @primaryKey fields 1:1", async () => {
     await GRAPHQL_CLIENT.query(
       `
       mutation CreateCTeam {
@@ -253,7 +253,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     await GRAPHQL_CLIENT.query(
@@ -265,7 +265,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     const queryResponse = await GRAPHQL_CLIENT.query(
@@ -287,19 +287,19 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
     expect(queryResponse.data.listCProjects).toBeDefined();
-    const items = queryResponse.data.listCProjects.items;
+    const { items } = queryResponse.data.listCProjects;
     expect(items.length).toEqual(1);
-    expect(items[0].projectId).toEqual('P1');
+    expect(items[0].projectId).toEqual("P1");
     expect(items[0].team).toBeDefined();
-    expect(items[0].team.teamId).toEqual('T1');
+    expect(items[0].team.teamId).toEqual("T1");
     expect(items[0].team.project).toBeDefined();
-    expect(items[0].team.project.projectId).toEqual('P1');
+    expect(items[0].team.project.projectId).toEqual("P1");
   });
 
-  test('Named connection 2 way navigation, with with custom @key fields 1:M', async () => {
+  test("Named connection 2 way navigation, with with custom @key fields 1:M", async () => {
     await GRAPHQL_CLIENT.query(
       `
       mutation CreateDProject {
@@ -309,7 +309,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     await GRAPHQL_CLIENT.query(
@@ -321,7 +321,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     await GRAPHQL_CLIENT.query(
@@ -333,7 +333,7 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
 
     const queryResponse = await GRAPHQL_CLIENT.query(
@@ -357,20 +357,20 @@ describe('@model with relational transformers', () => {
           }
       }
       `,
-      {},
+      {}
     );
     expect(queryResponse.data.listDProjects).toBeDefined();
-    const items = queryResponse.data.listDProjects.items;
+    const { items } = queryResponse.data.listDProjects;
     expect(items.length).toEqual(1);
-    expect(items[0].projectId).toEqual('P1');
+    expect(items[0].projectId).toEqual("P1");
     expect(items[0].teams).toBeDefined();
     expect(items[0].teams.items).toBeDefined();
     expect(items[0].teams.items.length).toEqual(2);
-    expect(items[0].teams.items[0].teamId).toEqual('T1');
+    expect(items[0].teams.items[0].teamId).toEqual("T1");
     expect(items[0].teams.items[0].project).toBeDefined();
-    expect(items[0].teams.items[0].project.projectId).toEqual('P1');
-    expect(items[0].teams.items[1].teamId).toEqual('T2');
+    expect(items[0].teams.items[0].project.projectId).toEqual("P1");
+    expect(items[0].teams.items[1].teamId).toEqual("T2");
     expect(items[0].teams.items[1].project).toBeDefined();
-    expect(items[0].teams.items[1].project.projectId).toEqual('P1');
+    expect(items[0].teams.items[1].project.projectId).toEqual("P1");
   });
 });