+++ packages/amplify-graphql-default-value-transformer/CHANGELOG.md
@@ -41,6 +41,20 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - migrate auth, maps-to, relational, default value transformer to CDK v2 ([#875](https://github.com/aws-amplify/amplify-category-api/issues/875)) ([5c714a9](https://github.com/aws-amplify/amplify-category-api/commit/5c714a9a8436be343477574cb5523c23c96c9338))
 
+## [0.7.7](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@0.7.6...@aws-amplify/graphql-default-value-transformer@0.7.7) (2023-01-26)
+
+**Note:** Version bump only for package @aws-amplify/graphql-default-value-transformer
+
+## [0.7.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@0.7.5...@aws-amplify/graphql-default-value-transformer@0.7.6) (2023-01-12)
+
+**Note:** Version bump only for package @aws-amplify/graphql-default-value-transformer
+
+## [0.7.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@0.7.4...@aws-amplify/graphql-default-value-transformer@0.7.5) (2023-01-12)
+
+### Bug Fixes
+
+- **api:** default value for model name start with lowercase ([df27e63](https://github.com/aws-amplify/amplify-category-api/commit/df27e63dd789c0156a94fdff3142acf65f460c65))
+
 ## [0.7.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@0.7.3...@aws-amplify/graphql-default-value-transformer@0.7.4) (2022-12-03)
 
 **Note:** Version bump only for package @aws-amplify/graphql-default-value-transformer