+++ packages/amplify-graphql-model-transformer/CHANGELOG.md
@@ -53,6 +53,18 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - migrate index and model transformers to CDK v2 ([#860](https://github.com/aws-amplify/amplify-category-api/issues/860)) ([886ab6c](https://github.com/aws-amplify/amplify-category-api/commit/886ab6c1eb699f9a09f273c76b3c419c73004f9b))
 
+## [0.16.7](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-model-transformer@0.16.6...@aws-amplify/graphql-model-transformer@0.16.7) (2023-01-26)
+
+**Note:** Version bump only for package @aws-amplify/graphql-model-transformer
+
+## [0.16.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-model-transformer@0.16.5...@aws-amplify/graphql-model-transformer@0.16.6) (2023-01-12)
+
+**Note:** Version bump only for package @aws-amplify/graphql-model-transformer
+
+## [0.16.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-model-transformer@0.16.4...@aws-amplify/graphql-model-transformer@0.16.5) (2023-01-12)
+
+**Note:** Version bump only for package @aws-amplify/graphql-model-transformer
+
 ## [0.16.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-model-transformer@0.16.3...@aws-amplify/graphql-model-transformer@0.16.4) (2022-12-03)
 
 **Note:** Version bump only for package @aws-amplify/graphql-model-transformer